feat: Add metadata to profil, plan, groupes
