+++ src/app/(layout-with-navigation)/(simulation)/groupes/creer/page.tsx
@@ -1,24 +1,22 @@
-'use client'
-
-import Meta from '@/components/misc/Meta'
 import GoBackLink from '@/design-system/inputs/GoBackLink'
 import Title from '@/design-system/layout/Title'
 import AutoCanonicalTag from '@/design-system/utils/AutoCanonicalTag'
-import { useClientTranslation } from '@/hooks/useClientTranslation'
+import { getServerTranslation } from '@/helpers/getServerTranslation'
+import { Metadata } from 'next'
 import GroupCreationForm from './_component/GroupCreationForm'
 
-export default function CreerGroupePage() {
-  const { t } = useClientTranslation()
+export const metadata: Metadata = {
+  title:
+    'Créer un groupe et calculer notre empreinte carbone - Nos Gestes Climat',
+  description:
+    "Calculez votre empreinte carbone en groupe et comparez la avec l'empreinte de vos proches grâce au simulateur de bilan carbone personnel Nos Gestes Climat.",
+}
+
+export default async function CreerGroupePage() {
+  const { t } = await getServerTranslation()
 
   return (
     <div className="p-4 md:p-8">
-      <Meta
-        title={t('Créer un groupe et calculer notre empreinte carbone')}
-        description={t(
-          "Calculez votre empreinte carbone en groupe et comparez la avec l'empreinte de vos proches grâce au simulateur de bilan carbone personnel Nos Gestes Climat."
-        )}
-      />
-
       <AutoCanonicalTag />
 
       <GoBackLink className="mb-4 font-bold" />
