+++ src/app/(layout-with-navigation)/(simulation)/groupes/supprimer/index.tsx
@@ -4,10 +4,17 @@ import Title from '@/design-system/layout/Title'
 import AutoCanonicalTag from '@/design-system/utils/AutoCanonicalTag'
 import { Member } from '@/types/groups'
 import { captureException } from '@sentry/react'
+import { Metadata } from 'next'
 import { useRouter } from 'next/navigation'
 import { useTranslation } from 'react-i18next'
 import { useFetchGroup } from '../_hooks/useFetchGroup'
 import { useDeleteGroup } from './_hooks/useDeleteGroup'
+
+export const metadata: Metadata = {
+  title: 'Supprimer mes données - Nos Gestes Climat',
+  description:
+    'Supprimez vos données de groupe enregistrées dans le simulateur Nos Gestes Climat.',
+}
 
 export default function SupprimerGroupePage({
   searchParams,
