+++ src/app/(layout-with-navigation)/(simulation)/profil/page.tsx
@@ -6,7 +6,7 @@ import { Metadata } from 'next'
 import ProfilPageContent from './_components/ProfilPageContent'
 
 export const metadata: Metadata = {
-  title: 'Mon profil',
+  title: 'Mon profil, voir mon empreinte carbone - Nos Gestes Climat',
   description:
     'Explorez et modifiez les informations que vous avez saisies dans le parcours nosgestesclimat.',
 }
