+++ src/constants/urls.ts
@@ -28,3 +28,8 @@ export const SAVE_SIMULATION_URL =
   (process.env.NODE_ENV === 'development'
     ? 'http://localhost:3001'
     : SERVER_URL) + '/email-simulation'
+
+export const APP_URL =
+  process.env.NODE_ENV === 'development'
+    ? 'http://localhost:3000'
+    : 'https://nosgestesclimat.fr'
